requirements.txt
openai
speechrecognition
python-dotenv
pyaudio
requests
pyttsx3
pywhatkit
elevenlabs 